+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## 1.0.0
+
+- Initial release
+- Support for simple mobile, tablet, and desktop responsive layouts
+- Includes:
+  - `ResponsiveChild`
+  - `ResponsiveLayout`
+  - `ResponsiveValue`
+  - `ResponsiveVisibility`
+  - `ResponsiveBuilder`
+  - `ResponsivePadding`
+  - `ResponsiveSizedBox`
+  - `ResponsiveSettingsProvider`
+- Allows custom breakpoints
+- Lightweight, zero dependencies beyond Flutter